+++ R/nlinla.R
@@ -14,13 +14,13 @@ bru_compute_linearisation <- function(...) {
 #' @param model A [bru_model] object
 #' @param lhood_expr A predictor expression
 #' @param data Input data
+#' @param input Precomputed component inputs from `evaluate_inputs()`
 #' @param state The state information, as a list of named vectors
-#' @param A A-matrix information:
-#' * For `bru_component`: Precomputed A-matrix for the component
-#' * For `bru_like`: A list of named A-matrices for the components in the
-#' likelihood for the component
-#' * For `bru_like_list`: A list, where each element is a list of named
-#' A-matrices.
+#' @param comp_simple Component evaluation information
+#' * For `bru_component`: `bru_mapper_taylor` object
+#' * For `bru_like`: A `comp_simple_list` object
+#'   for the components in the likelihood
+#' * For `bru_like_list`: A `comp_simple_list_list` object
 #' @param effects
 #' * For `bru_component`:
 #' Precomputed effect list for all components involved in the likelihood
@@ -36,8 +36,9 @@ bru_compute_linearisation.component <- function(cmp,
                                                 model,
                                                 lhood_expr,
                                                 data,
+                                                input,
                                                 state,
-                                                A,
+                                                comp_simple,
                                                 effects,
                                                 pred0,
                                                 allow_latent,
@@ -45,6 +46,13 @@ bru_compute_linearisation.component <- function(cmp,
                                                 eps,
                                                 ...) {
   label <- cmp[["label"]]
+
+  if (!inherits(comp_simple, "bru_mapper_taylor")) {
+    warning("Non-linear component mappers not fully supported!", immediate. = TRUE)
+  }
+  A <- ibm_jacobian(comp_simple,
+                    input = input[[label]],
+                    state = state[[label]])
 
   assume_rowwise <- !allow_latent && !allow_combine && is.data.frame(data)
 
@@ -59,8 +67,8 @@ bru_compute_linearisation.component <- function(cmp,
     }
   }
 
-  if (identical(cmp[["main"]][["type"]], "offset")) {
-    # Zero-column matrix, since an offset has no latent state variables.
+  if (cmp[["main"]][["type"]] %in% c("offset", "const")) {
+    # Zero-column matrix, since a const/offset has no latent state variables.
     return(Matrix::sparseMatrix(
       i = c(),
       j = c(),
@@ -69,7 +77,6 @@ bru_compute_linearisation.component <- function(cmp,
     ))
   }
 
-  label <- cmp[["label"]]
   triplets <- list(
     i = integer(0),
     j = integer(0),
@@ -77,7 +84,7 @@ bru_compute_linearisation.component <- function(cmp,
   )
   for (k in seq_len(NROW(state[[label]]))) {
     row_subset <- which(A[, k] != 0.0)
-    if (allow_latent || (length(row_subset) > 0)) {
+    if (length(row_subset) > 0) {
       state_eps <- state
       state_eps[[label]][k] <- state[[label]][k] + eps
       # TODO:
@@ -152,8 +159,9 @@ bru_compute_linearisation.component <- function(cmp,
 bru_compute_linearisation.bru_like <- function(lhood,
                                                model,
                                                data,
+                                               input,
                                                state,
-                                               A,
+                                               comp_simple,
                                                eps,
                                                ...) {
   allow_latent <- lhood[["allow_latent"]]
@@ -163,11 +171,10 @@ bru_compute_linearisation.bru_like <- function(lhood,
     lhood[["include_components"]],
     lhood[["exclude_components"]]
   )
-  effects <- evaluate_effect_single(
-    model[["effects"]][included],
-    state = state,
-    data = NULL,
-    A = A
+  effects <- evaluate_effect_single_state(
+    comp_simple[included],
+    input = input[included],
+    state = state[included],
   )
 
   lhood_expr <- bru_like_expr(lhood, model[["effects"]])
@@ -185,34 +192,37 @@ bru_compute_linearisation.bru_like <- function(lhood,
     # and possibly expand to full size
     if (length(pred0) == 1) {
       if (is.data.frame(data) ||
-        inherits(data, c(
-          "SpatialPointsDataFrame",
-          "SpatialPolygonsDataFrame",
-          "SpatialLinesDataFrame"
-        ))) {
+          inherits(data, c(
+            "SpatialPointsDataFrame",
+            "SpatialPolygonsDataFrame",
+            "SpatialLinesDataFrame"
+          ))) {
         pred0 <- rep(pred0, NROW(data))
       }
     }
   }
 
-  # Compute derivatives for each non-offset component
+  # Compute derivatives for each non-const/offset component
   B <- list()
   offset <- pred0
   # Either this loop or the internal bru_component specific loop
   # can in principle be parallelised.
   for (label in included) {
-    if (!identical(model[["effects"]][[label]][["main"]][["type"]], "offset")) {
+    if (ibm_n(model[["effects"]][[label]][["mapper"]]) > 0) {
       if (lhood[["linear"]] && !lhood[["allow_combine"]]) {
         # If linear and no combinations allowed, just need to copy the
         # non-offset A matrix, and possibly expand to full size
-        if (NROW(A[[label]]) == 1) {
+        A <- ibm_jacobian(comp_simple[[label]],
+                          input[[label]],
+                          state[[label]])
+        if (NROW(A) == 1) {
           if (NROW(offset) > 1) {
-            B[[label]] <- Matrix::kronecker(rep(1, NROW(offset)), A[[label]])
+            B[[label]] <- Matrix::kronecker(rep(1, NROW(offset)), A)
           } else {
-            B[[label]] <- A[[label]]
+            B[[label]] <- A
           }
         } else {
-          B[[label]] <- A[[label]]
+          B[[label]] <- A
         }
       } else {
         B[[label]] <-
@@ -221,8 +231,9 @@ bru_compute_linearisation.bru_like <- function(lhood,
             model = model,
             lhood_expr = lhood_expr,
             data = data,
+            input = input,
             state = state,
-            A = A[[label]],
+            comp_simple = comp_simple[[label]],
             effects = effects,
             pred0 = pred0,
             allow_latent = lhood[["allow_latent"]],
@@ -235,7 +246,7 @@ bru_compute_linearisation.bru_like <- function(lhood,
     }
   }
 
-  list(A = B, offset = offset)
+  bru_mapper_taylor(offset = offset, jacobian = B, state0 = NULL)
 }
 
 #' @param lhoods A `bru_like_list` object
@@ -244,8 +255,9 @@ bru_compute_linearisation.bru_like <- function(lhood,
 #' @rdname bru_compute_linearisation
 bru_compute_linearisation.bru_like_list <- function(lhoods,
                                                     model,
+                                                    input,
                                                     state,
-                                                    A,
+                                                    comp_simple,
                                                     eps = 1e-5, # TODO: set more intelligently
                                                     ...) {
   lapply(seq_along(lhoods), function(idx) {
@@ -254,8 +266,9 @@ bru_compute_linearisation.bru_like_list <- function(lhoods,
       x,
       model = model,
       data = x[["data"]],
+      input = input[[idx]],
       state = state,
-      A = A[[idx]],
+      comp_simple = comp_simple[[idx]],
       eps = eps,
       ...
     )
@@ -264,6 +277,10 @@ bru_compute_linearisation.bru_like_list <- function(lhoods,
 
 #' @export
 #' @rdname bru_compute_linearisation
-bru_compute_linearisation.bru_model <- function(model, lhoods, state, A, ...) {
-  bru_compute_linearisation(lhoods, model = model, state = state, A = A, ...)
+bru_compute_linearisation.bru_model <- function(model, lhoods,
+                                                input, state,
+                                                comp_simple, ...) {
+  bru_compute_linearisation(lhoods, model = model,
+                            input = input, state = state,
+                            comp_simple = comp_simple, ...)
 }