+++ NEWS.md
@@ -11,9 +11,30 @@
   weights, passed on to the `INLA::inla()` weights argument. Evaluated in the
   data context.
   
+* The `<component>_eval()` methods available in predictor expressions
+  now handle optional scaling weights, like in ordinary component effect
+  evaluation.
+
 * New vignettes on the `bru_mapper` system, `component` definitions,
   and `prediction_scores`
-
+  
+* General overhaul of the `bru_mapper` and linearised predictor system,
+  to prepare for new features.
+  
+  * Add `ibm_eval` generic for evaluating mappers for given states.
+  
+  * Add `bru_mapper_taylor`, used as an internal mapper for linearised
+  mappers. This and `ibm_eval` is aimed at future support for nonlinear
+  mappers. Associated new generic methods: `ibm_{is_linear,jacobian,linear}`.
+  
+  * New mapper implementations should use `ibm_jacobian` instead of `ibm_amatrix`.
+    This allows defining a linearised mapper via
+   `ibm_eval(input, state0) + ibm_jacobian(input, state0) %*% (state - state0)`.
+  
+  * New mapper class `bru_mapper_const`, which replaces `bru_mapper_offset`.
+  `bru_mapper_offset` is now deprecated and will produce warnings.
+  
+  
 ## Bug fixes
 
 * Enable both datum/ensemble container for ellipsoid information, to support